# [Account plugin] add cache to authentication method