--- a/kinto/plugins/accounts/__init__.py
+++ b/kinto/plugins/accounts/__init__.py
@@ -1,6 +1,8 @@
 from kinto.authorization import PERMISSIONS_INHERITANCE_TREE
 from pyramid.exceptions import ConfigurationError
 
+ACCOUNT_CACHE_KEY = 'accounts:{}:verified'
+
 
 def includeme(config):
     config.add_api_capability(