--- a/kinto/plugins/accounts/authentication.py
+++ b/kinto/plugins/accounts/authentication.py
@@ -4,11 +4,14 @@ from pyramid import authentication as base_auth
 from kinto.core import utils
 from kinto.core.storage import exceptions as storage_exceptions
 
+from . import ACCOUNT_CACHE_KEY
+
 
 def account_check(username, password, request):
     settings = request.registry.settings
     hmac_secret = settings['userid_hmac_secret']
-    cache_key = utils.hmac_digest(hmac_secret, '{}:{}'.format(username, password))
+    cache_key = utils.hmac_digest(hmac_secret, ACCOUNT_CACHE_KEY.format(username))
+    cache_ttl = settings.get('account_cache_ttl_seconds', 30)
 
     # Check cache to see whether somebody has recently logged in with the same
     # username and password.
@@ -16,7 +19,8 @@ def account_check(username, password, request):
     cache_result = cache.get(cache_key)
 
     # Username and password is correct. No need to compare hashes
-    if cache_result == "1":
+    if cache_result:
+        cache.expire(cache_key, cache_ttl)
         return True
 
     # Back to standard procedure
@@ -31,10 +35,8 @@ def account_check(username, password, request):
     hashed = existing['password'].encode(encoding='utf-8')
     pwd_str = password.encode(encoding='utf-8')
     # Check if password is valid (it is a very expensive computation)
-    if bcrypt.checkpw(pwd_str, hashed):  # Match!
-        # Remember result so we don't have to calculate bcrypt hash next time.
-        cache.set(cache_key, "1", ttl=30)  # Time to live: 30 seconds
-        # Return anything but None.
+    if bcrypt.checkpw(pwd_str, hashed):
+        cache.set(cache_key, "1", ttl=cache_ttl)
         return True