When debugging my application, I've noticed very poor performance when logged in using account plugin. It turns out that the biggest bottle neck is bcrypt and its `hashpw` method (https://github.com/Kinto/kinto/blob/master/kinto/plugins/accounts/authentication.py#L18). Making sure it is executed no more than once per request drastically reduces response times.

### Before:
```
> ab -n 100 -A admin:admin 127.0.0.1:8888/v1/
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient).....done


Server Software:        waitress
Server Hostname:        127.0.0.1
Server Port:            8888

Document Path:          /v1/
Document Length:        697 bytes

Concurrency Level:      1
Time taken for tests:   39.127 seconds
Complete requests:      100
Failed requests:        0
Total transferred:      95200 bytes
HTML transferred:       69700 bytes
Requests per second:    2.56 [#/sec] (mean)
Time per request:       391.272 [ms] (mean)
Time per request:       391.272 [ms] (mean, across all concurrent requests)
Transfer rate:          2.38 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:   387  391   4.8    390     427
Waiting:      386  391   4.8    390     427
Total:        387  391   4.8    390     427

Percentage of the requests served within a certain time (ms)
  50%    390
  66%    391
  75%    392
  80%    393
  90%    395
  95%    399
  98%    402
  99%    427
 100%    427 (longest request)
```
![image](https://user-images.githubusercontent.com/41941/33650408-2ad2cec2-da62-11e7-942f-14a7168be1b8.png)

### After:
```
> ab -n 100 -A admin:admin 127.0.0.1:8888/v1/
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient).....done


Server Software:        waitress
Server Hostname:        127.0.0.1
Server Port:            8888

Document Path:          /v1/
Document Length:        697 bytes

Concurrency Level:      1
Time taken for tests:   0.392 seconds
Complete requests:      100
Failed requests:        0
Total transferred:      95200 bytes
HTML transferred:       69700 bytes
Requests per second:    255.05 [#/sec] (mean)
Time per request:       3.921 [ms] (mean)
Time per request:       3.921 [ms] (mean, across all concurrent requests)
Transfer rate:          237.11 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:     2    4   4.1      3      43
Waiting:        2    4   4.1      3      43
Total:          2    4   4.1      3      43

Percentage of the requests served within a certain time (ms)
  50%      3
  66%      4
  75%      4
  80%      4
  90%      5
  95%      5
  98%      6
  99%     43
 100%     43 (longest request)
```

![screenshot_20171206_084711](https://user-images.githubusercontent.com/41941/33650434-42cae8ac-da62-11e7-825d-592b32693369.png)

- [ ] Add documentation.
- [ ] Add tests.
- [ ] Add a changelog entry.
- [ ] Add your name in the contributors file.
- [ ] If you changed the HTTP API, update the [API_VERSION](https://github.com/Kinto/kinto/blob/master/kinto/__init__.py#L15) constant and add an API changelog entry [in the docs](https://github.com/Kinto/kinto/blob/master/docs/api/index.rst)
- [ ] If you added a new configuration setting, update the [`kinto.tpl`](https://github.com/Kinto/kinto/blob/master/kinto/config/kinto.tpl) file with it.