@Stanley 2017-12-07
Done. All tests are passing :) Anything else?