@Natim 2017-12-07
Can you add a line in the CHANGELOG?