@Natim 2017-12-12
> if an attacker gain access the cache layer and the HMAC secret, than we can have a password leakage.

No because it is a hmac and it is not reversible, however we can rotate the hmac key with this plugin since we can wipe the cache whenever we want.