@Natim 2017-12-13
> The problem isn't about being reversible, but being deterministic.

Oh right, we should use the username hash of the username and then have a random salt and a saltedhashed password as the value.

Thanks for detecting this!