@leplatrem 2018-02-02
We can merge once the following is done:

- [ ] unit tests that assert a cache key is set (or bcrypt is not called on second attempt)
- [ ] unit tests that assert what happens when user/pass is changed/deleted
- [ ] the TTL value in settings
- [ ] a cache refresh on each request (basically call `cache.set()`)

@Stanley are you still interested in working on this and go through those last steps? Let us know, maybe somebody can help and take over!

Thanks!